+++ challenges/CROMU_00088/lib/filesystem.c
@@ -27,6 +27,7 @@ THE SOFTWARE.
 #include "filesystem.h"
 
 #include <libcgc.h>
+#include <stdint.h>
 
 #include "malloc.h"
 #include "stdio.h"
@@ -677,7 +678,7 @@ int readMemoryFile(fileHandleType fh, char *buffer, unsigned int size,
   unsigned int copySize;
 
   struct memoryFileInfo {
-    unsigned int address;
+    uintptr_t address;
     unsigned int size;
     char accessType;
 
@@ -727,7 +728,7 @@ int writeMemoryFile(fileHandleType fh, char *buffer, unsigned int size) {
   unsigned int copySize;
 
   struct memoryFileInfo {
-    unsigned int address;
+    uintptr_t address;
     unsigned int size;
     char accessType;
 
@@ -895,7 +896,7 @@ int deleteFile(fileHandleType fh, securityIdType securityID) {
   int retval;
 
   struct memoryFileInfo {
-    unsigned int address;
+    uintptr_t address;
     unsigned int size;
     char accessType;
