+++ challenges/KPRCA_00001/lib/stdlib.h
@@ -4,12 +4,13 @@
 #include <libcgc.h>
 #include <stdarg.h>
 #include <stddef.h>
+#include <stdint.h>
 
 extern int writeall(int fd, const char *buf, size_t n);
 extern int readuntil(int fd, char *buf, size_t n, char delim);
 
 extern int debug_printf(const char *fmt, ...);
-extern int fdprintf(int fd, const char *fmt, ...);
+extern int fdprintf(intptr_t fd, const char *fmt, ...);
 extern int sprintf(char *s, const char *fmt, ...);
 extern int snprintf(char *s, size_t size, const char *fmt, ...);
 extern int vsnprintf(char *s, size_t size, const char *fmt, va_list ap);
